reorganizing types since they'll grow in size
