+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export type { uWaveAPI } from './api';
+export type { Commands, SocketEvents, SocketPayloadsMap } from './socket';
+export type { SourceData } from './sources';
